Split projects lib file to multiple separate file
