+++ lib/pages/index.js
@@ -1,0 +1,11 @@
+import * as OverviewPage from "./overview";
+import ProjectPage from "./project";
+import * as SettingsPage from "./settings";
+
+const defaults = {
+    overview: OverviewPage,
+    project: ProjectPage,
+    settings: SettingsPage,
+};
+
+export default defaults;
