+++ lib/pages/project/index.js
@@ -1,0 +1,11 @@
+import * as Add from "./add";
+import * as Main from "./main";
+import * as Page from "./page";
+
+const defaults = {
+    add: Add,
+    main: Main,
+    page: Page,
+};
+
+export default defaults;
